Move types/functions around
